Modifications de plants controller

+++ src/main/java/com/epf/persistence/mapper/PlantsMapper.java
@@ -2,9 +2,9 @@ package com.epf.persistence.mapper;
 
 import com.epf.persistence.model.Plants;
 import org.springframework.jdbc.core.RowMapper;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
+
 
 public class PlantsMapper implements RowMapper<Plants> {
     @Override
@@ -13,10 +13,10 @@ public class PlantsMapper implements RowMapper<Plants> {
                 rs.getInt("id_plante"),
                 rs.getString("nom"),
                 rs.getInt("point_de_vie"),
-                rs.getInt("attaque_par_seconde"),
+                rs.getBigDecimal("attaque_par_seconde"),  // Changed from getInt
                 rs.getInt("degat_attaque"),
                 rs.getInt("cout"),
-                rs.getInt("soleil_par_seconde"),
+                rs.getBigDecimal("soleil_par_seconde"),   // Changed from getInt
                 rs.getString("effet"),
                 rs.getString("chemin_image")
         );
